requirements.txt
# OnShelf AI Agent System Requirements

# Core ML/AI Libraries
instructor>=0.4.0
openai>=1.3.0
anthropic>=0.8.0
google-generativeai>=0.3.0
pydantic>=2.5.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0

# Image Processing
pillow>=10.0.0

# Web Framework & API
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# Dashboard & Visualization
streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0

# Async Support
asyncio>=3.4.3
aiofiles>=23.0.0

# Environment & Config
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
mypy>=1.7.0

# Monitoring & Logging
python-json-logger>=2.0.0

# Optional Performance
redis>=5.0.0  # For caching
numpy>=1.24.0  # For numerical operations 